#536: Extracted interface `ImporterFactory` from abstract base class that is now named`AbstractImporterFactory`.
