+++ doc/plugin_developer_guide.md
@@ -13,10 +13,12 @@ This guide describes how to develop [plugins](plugins.md) for OpenFastTrace (OFT
    </dependency>
    ```
 
-2. Create a new class (e.g. `com.example.oft.import.MyImporter`) implementing one of the following interfaces:
+2. Create a new class (e.g. `com.example.oft.import.MyImporter`) using one of the following plugin APIs:
    * [`org.itsallcode.openfasttrace.api.report.ReporterFactory`](https://github.com/itsallcode/openfasttrace/blob/main/api/src/main/java/org/itsallcode/openfasttrace/api/report/ReporterFactory.java): Generate tracing report
    * [`org.itsallcode.openfasttrace.api.importer.ImporterFactory`](https://github.com/itsallcode/openfasttrace/blob/main/api/src/main/java/org/itsallcode/openfasttrace/api/importer/ImporterFactory.java): Import requirements from a new file format
    * [`org.itsallcode.openfasttrace.api.exporter.ExporterFactory`](https://github.com/itsallcode/openfasttrace/blob/main/api/src/main/java/org/itsallcode/openfasttrace/api/exporter/ExporterFactory.java): Export requirements in a new file format
+
+   Importer plugins will usually extend [`org.itsallcode.openfasttrace.api.importer.AbstractImporterFactory`](https://github.com/itsallcode/openfasttrace/blob/main/api/src/main/java/org/itsallcode/openfasttrace/api/importer/AbstractImporterFactory.java) instead of implementing `ImporterFactory` directly so they inherit the standard context handling.
 
 3. Create a file in `src/main/resources/$INTERFACE_FQN`, using the fully qualified class name of the interface as file name.
